test(rstest): focus adapter tests on the @effect/vitest suite
